MAGMA-105:  - Mapped DecimalType to SQL type DOUBLE.

+++ datasource-jdbc/src/main/java/org/obiba/magma/datasource/jdbc/SqlTypes.java
@@ -96,7 +96,7 @@ class SqlTypes {
       return "INTEGER";
     }
     if(valueType.equals(DecimalType.get())) {
-      return "DECIMAL";
+      return "DOUBLE";
     }
     // if(valueType.equals(DateType.get())) {
     // return "DATE";
